Update libfsfat to version 20220925 / rev 2 via SR 1038466
https://build.opensuse.org/request/show/1038466
by user jengelh + dimstar_suse
delete system-libs.patch

+++ packages/libf/libfsfat/libfsfat.spec
@@ -27,11 +27,8 @@ URL:            https://github.com/libyal/libfsfat
 Source:         https://github.com/libyal/libfsfat/releases/download/%version/libfsfat-experimental-%version.tar.gz
 Source2:        https://github.com/libyal/libfsfat/releases/download/%version/libfsfat-experimental-%version.tar.gz.asc
 Source3:        %name.keyring
-Patch1:         system-libs.patch
 BuildRequires:  %python_module devel
 BuildRequires:  c_compiler
-BuildRequires:  gettext-tools >= 0.21
-BuildRequires:  libtool
 BuildRequires:  pkg-config
 BuildRequires:  python-rpm-macros
 BuildRequires:  pkgconfig(libbfio) >= 20220120
@@ -45,13 +42,14 @@ BuildRequires:  pkgconfig(libcsplit) >= 20220109
 BuildRequires:  pkgconfig(libcthreads) >= 20220102
 BuildRequires:  pkgconfig(libfcache) >= 20220110
 BuildRequires:  pkgconfig(libfcrypto) >= 20210415
-BuildRequires:  pkgconfig(libfdata) >= 20211023
+BuildRequires:  pkgconfig(libfdata) >= 20220111
 BuildRequires:  pkgconfig(libfdatetime) >= 20220112
 BuildRequires:  pkgconfig(libfguid) >= 20220113
 BuildRequires:  pkgconfig(libhmac) >= 20220425
 BuildRequires:  pkgconfig(libuna) >= 20220611
 BuildRequires:  pkgconfig(python3)
 %python_subpackages
+# Various notes: https://en.opensuse.org/libyal
 
 %description
 libfsfat is a library to access the File Allocation Table (FAT) file
@@ -93,22 +91,19 @@ reading FAT File Systems.
 %autosetup -p1
 
 %build
-autoreconf -fi
-# OOT builds are presently broken, so we have to install
-# within each python iteration now, not in %%install.
 %{python_expand #
-# see libcdata for version-sc
 echo "V_%version { global: *; };" >v.sym
 %configure --disable-static --enable-wide-character-type --enable-python \
 	PYTHON_VERSION="%{$python_bin_suffix}" \
 	LDFLAGS="-Wl,--version-script=$PWD/v.sym"
+grep ' '' ''local' config.log && exit 1
 %make_build
 %make_install DESTDIR="%_builddir/rt"
 %make_build clean
 }
 
 %install
-mv %_builddir/rt/* %buildroot/
+mv "%_builddir/rt"/* %buildroot/
 find "%buildroot" -type f -name "*.la" -delete -print
 
 %post   -n %lname -p /sbin/ldconfig
